refactor: split sync_test.go into focused test files
Split the 2,487-line sync_test.go (36 tests) into 7 focused files:
- sync_test_helpers_test.go: shared helpers, legacy DDL, syncTestHelper
- sync_state_test.go: sync state and worker lifecycle (6 tests)
- sync_backfill_test.go: machine ID and repo identity backfill (6 tests)
- sync_identity_test.go: repo identity CRUD and schema migrations (8 tests)
- sync_queries_test.go: sync query functions and timestamps (5 tests)
- sync_ordering_test.go: sync ordering and dependency enforcement (8 tests)
- sync_remap_test.go: patch ID round-trip and job remapping (4 tests)

No test logic changes. Follows the pattern set by review_test.go split (#379).
